Module aws_sdk_emrcontainers::model

Modules

cloud_watch_monitoring_configuration

See CloudWatchMonitoringConfiguration

configuration

See Configuration

configuration_overrides

See ConfigurationOverrides

container_provider

See ContainerProvider

eks_info

See EksInfo

endpoint

See Endpoint

job_driver

See JobDriver

job_run

See JobRun

monitoring_configuration

See MonitoringConfiguration

s3_monitoring_configuration

See S3MonitoringConfiguration

spark_submit_job_driver

See SparkSubmitJobDriver

virtual_cluster

See VirtualCluster

Structs

CloudWatchMonitoringConfiguration

A configuration for CloudWatch monitoring. You can configure your jobs to send log information to CloudWatch Logs.

Configuration

A configuration specification to be used when provisioning virtual clusters, which can include configurations for applications and software bundled with Amazon EMR on EKS. A configuration consists of a classification, properties, and optional nested configurations. A classification refers to an application-specific configuration file. Properties are the settings you want to change in that file.

ConfigurationOverrides

A configuration specification to be used to override existing configurations.

ContainerProvider

The information about the container provider.

EksInfo

The information about the EKS cluster.

Endpoint

This entity represents the endpoint that is managed by Amazon EMR on EKS.

JobDriver

Specify the driver that the job runs on.

JobRun

This entity describes a job run. A job run is a unit of work, such as a Spark jar, PySpark script, or SparkSQL query, that you submit to Amazon EMR on EKS.

MonitoringConfiguration

Configuration setting for monitoring.

S3MonitoringConfiguration

Amazon S3 configuration for monitoring log publishing. You can configure your jobs to send log information to Amazon S3.

SparkSubmitJobDriver

The information about job driver for Spark submit.

VirtualCluster

This entity describes a virtual cluster. A virtual cluster is a Kubernetes namespace that Amazon EMR is registered with. Amazon EMR uses virtual clusters to run jobs and host endpoints. Multiple virtual clusters can be backed by the same physical cluster. However, each virtual cluster maps to one namespace on an EKS cluster. Virtual clusters do not create any active resources that contribute to your bill or that require lifecycle management outside the service.

Enums

ContainerInfo

ContainerProviderType

EndpointState

FailureReason

JobRunState

PersistentAppUi

VirtualClusterState